Create the validation as part of a callback
