--- projects/cloud/config/credentials/environment.yml.enc
@@ -1,1 +1,0 @@
-ykOiBj0+HnRnvwiaTy3bZXHbi4rfPpXNBSuw1zi3C9vgmKdcq0+VUjcQTLrphw+1t41+9Nz5WhRYvqfVhIbDaPiP7as=--wycAJd3Qz7V3g0Ty--yJR4Xbgw54DPn+awvrbNkw==
